Refactor: Move code in App.js into separate file
